+++ NEWS.md
@@ -9,3 +9,8 @@
 # geess 0.1.2
 
 * Revised in accordance with the comments from CRAN
+
+# geess 1.0.0
+
+* Revised geess function
+* Added predict.geess function